Show list of own encounters at index
To see them displayed, one must have associated
their `Account` with their `user`;
currently, the only way to do this is from the
Admin screen, only accessible to `is_staff` users;
but it should be in the profile, along with the
API key setting and other such.

Also TODO, show some feedback on upload progress.